memperbarui KtpRepository dengan menyesuaikan tipe primary key menjadi Long pada JpaRepository. Perubahan ini dilakukan agar tipe data ID pada repository konsisten dengan tipe data yang digunakan pada entity dan layer service dalam pengelolaan data KTP

+++ src/main/java/com/example/ktp/repository/KtpRepository.java
@@ -3,6 +3,6 @@ package com.example.ktp.repository;
 import com.example.ktp.entity.Ktp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KtpRepository extends JpaRepository<Ktp, Integer> {
+public interface KtpRepository extends JpaRepository<Ktp, Long> {
 
 }
